+++ src/commands/dataconnect-sdk-generate.ts
@@ -5,10 +5,15 @@ import { Options } from "../options";
 import { DataConnectEmulator } from "../emulator/dataconnectEmulator";
 import { needProjectId } from "../projectUtils";
 import { loadAll } from "../dataconnect/load";
-import { logger } from "../logger";
 import { getProjectDefaultAccount } from "../auth";
-import { logLabeledSuccess } from "../utils";
+import { logBullet, logLabeledSuccess, logWarning } from "../utils";
 import { ServiceInfo } from "../dataconnect/types";
+import { Config } from "../config";
+import { Setup } from "../init";
+import * as dataconnectInit from "../init/features/dataconnect";
+import * as dataconnectSdkInit from "../init/features/dataconnect/sdk";
+import { FirebaseError } from "../error";
+import { postInitSaves } from "./init";
 
 type GenerateOptions = Options & { watch?: boolean };
 
@@ -21,42 +26,98 @@ export const command = new Command("dataconnect:sdk:generate")
   .action(async (options: GenerateOptions) => {
     const projectId = needProjectId(options);
 
-    const serviceInfos = await loadAll(projectId, options.config);
-    const serviceInfosWithSDKs = serviceInfos.filter((serviceInfo) =>
-      serviceInfo.connectorInfo.some((c) => {
-        return (
-          c.connectorYaml.generate?.javascriptSdk ||
-          c.connectorYaml.generate?.kotlinSdk ||
-          c.connectorYaml.generate?.swiftSdk ||
-          c.connectorYaml.generate?.dartSdk
+    let justRanInit = false;
+    let config = options.config;
+    if (!config || !config.has("dataconnect")) {
+      if (options.nonInteractive) {
+        throw new FirebaseError(
+          `No dataconnect project directory found. Please run ${clc.bold("firebase init dataconnect")} to set it up first.`,
         );
-      }),
-    );
-    if (!serviceInfosWithSDKs.length) {
-      logger.warn("No generated SDKs have been declared in connector.yaml files.");
-      logger.warn(`Run ${clc.bold("firebase init dataconnect:sdk")} to configure a generated SDK.`);
-      logger.warn(
-        `See https://firebase.google.com/docs/data-connect/web-sdk for more details of how to configure generated SDKs.`,
-      );
-      return;
-    }
-    async function generateSDK(serviceInfo: ServiceInfo): Promise<void> {
-      return DataConnectEmulator.generate({
-        configDir: serviceInfo.sourceDirectory,
-        watch: options.watch,
-        account: getProjectDefaultAccount(options.projectRoot),
-      });
-    }
-    if (options.watch) {
-      await Promise.race(serviceInfosWithSDKs.map(generateSDK));
-    } else {
-      for (const s of serviceInfosWithSDKs) {
-        await generateSDK(s);
       }
-      const services = serviceInfosWithSDKs.map((s) => s.dataConnectYaml.serviceId).join(", ");
-      logLabeledSuccess(
-        "dataconnect",
-        `Successfully Generated SDKs for services: ${clc.bold(services)}`,
+      logWarning("No dataconnect project directory found.");
+      logBullet(
+        `Running ${clc.bold("firebase init dataconnect")} to setup a dataconnect project directory.`,
       );
+      if (!config) {
+        const cwd = options.cwd || process.cwd();
+        config = new Config({}, { projectDir: cwd, cwd: cwd });
+      }
+      const setup: Setup = {
+        config: config.src,
+        rcfile: options.rc.data,
+        instructions: [],
+      };
+      await dataconnectInit.askQuestions(setup);
+      await dataconnectInit.actuate(setup, config, options);
+      await postInitSaves(setup, config);
+      justRanInit = true;
+      options.config = config;
     }
+
+    let serviceInfosWithSDKs = await loadAllWithSDKs(projectId, config);
+    if (!serviceInfosWithSDKs.length) {
+      if (justRanInit || options.nonInteractive) {
+        throw new FirebaseError(
+          `No generated SDKs are configured during init. Please run ${clc.bold("firebase init dataconnect:sdk")} to configure a generated SDK.`,
+        );
+      }
+      logWarning("No generated SDKs have been configured.");
+      logBullet(
+        `Running ${clc.bold("firebase init dataconnect:sdk")} to configure a generated SDK.`,
+      );
+      const setup: Setup = {
+        config: config.src,
+        rcfile: options.rc.data,
+        instructions: [],
+      };
+      await dataconnectSdkInit.askQuestions(setup);
+      await dataconnectSdkInit.actuate(setup, config);
+      justRanInit = true;
+      serviceInfosWithSDKs = await loadAllWithSDKs(projectId, config);
+    }
+
+    await generateSDKsInAll(options, serviceInfosWithSDKs, justRanInit);
   });
+
+async function loadAllWithSDKs(projectId: string, config: Config): Promise<ServiceInfo[]> {
+  const serviceInfos = await loadAll(projectId, config);
+  return serviceInfos.filter((serviceInfo) =>
+    serviceInfo.connectorInfo.some((c) => {
+      return (
+        c.connectorYaml.generate?.javascriptSdk ||
+        c.connectorYaml.generate?.kotlinSdk ||
+        c.connectorYaml.generate?.swiftSdk ||
+        c.connectorYaml.generate?.dartSdk
+      );
+    }),
+  );
+}
+
+async function generateSDKsInAll(
+  options: GenerateOptions,
+  serviceInfosWithSDKs: ServiceInfo[],
+  justRanInit: boolean,
+): Promise<void> {
+  async function generateSDK(serviceInfo: ServiceInfo): Promise<void> {
+    return DataConnectEmulator.generate({
+      configDir: serviceInfo.sourceDirectory,
+      watch: options.watch,
+      account: getProjectDefaultAccount(options.projectRoot),
+    });
+  }
+  if (options.watch) {
+    await Promise.race(serviceInfosWithSDKs.map(generateSDK));
+  } else {
+    if (justRanInit) {
+      return; // SDKs are already generated during init
+    }
+    for (const s of serviceInfosWithSDKs) {
+      await generateSDK(s);
+    }
+    const services = serviceInfosWithSDKs.map((s) => s.dataConnectYaml.serviceId).join(", ");
+    logLabeledSuccess(
+      "dataconnect",
+      `Successfully Generated SDKs for services: ${clc.bold(services)}`,
+    );
+  }
+}